Move ISampleStore to correct namespace
