+++ osu.Framework/Audio/Sample/ISampleChannel.cs
@@ -1,8 +1,6 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Framework.Audio.Track;
-
 namespace osu.Framework.Audio.Sample
 {
     /// <summary>
