+++ osu.Framework/Audio/Sample/ISampleStore.cs
@@ -1,9 +1,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Framework.Audio.Sample;
-
-namespace osu.Framework.Audio.Track
+namespace osu.Framework.Audio.Sample
 {
     public interface ISampleStore : IAdjustableResourceStore<SampleChannel>
     {
